requirements.txt
alembic==1.4.2
astroid==2.4.2
autopep8==1.5.3
Babel==2.8.0
blinker==1.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dnspython==1.16.0
docker==4.2.0
dockerpty==0.4.1
docopt==0.6.2
email-validator==1.1.1
flake8==3.8.3
Flask==1.1.2
Flask-Admin==1.5.6
Flask-BabelEx==0.9.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Principal==0.4.0
Flask-Script==2.0.6
Flask-Security==3.0.0
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
gitdb==4.0.5
GitPython==3.1.0
idna==2.9
importlib-metadata==1.6.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
mypy==0.781
mypy-extensions==0.4.3
mysql-connector-python==8.0.20
passlib==1.7.2
pkg-resources==0.0.0
protobuf==3.12.2
pycodestyle==2.6.0
pyflakes==2.2.0
pykwalify==1.7.0
python-box==4.2.3
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
semver==2.10.2
six==1.15.0
smmap==3.0.4
speaklater==1.3
spython==0.0.79
SQLAlchemy==1.3.17
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.25.9
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.1
zipp==3.1.0
